commit 1

+++ src/main/java/com/course_sys/entity/User.java
@@ -91,6 +91,6 @@ public class User implements UserDetails, Comparable<User> {
 
     @Override
     public int compareTo(User o) {
-        return (int) (this.id - o.id);
+        return this.id - o.id;
     }
 }
